Address review: extract shared file resolution, remove unused FileName, fix docs and switch order
- Extract ResolveFileReferences helper into FileUploadStore to eliminate
  near-verbatim duplicate logic between DashboardServiceData and
  PipelineActivityReporter (including shared FileReference type)
- Remove unused FileName property from PublishingPromptInputAnswer
- Fix MaxFileSize XML doc to reflect actual 100 MB server default
- Fix switch arm ordering in Publishers AppHost so publish-test is
  reachable when publisher is 'default'
