+++ src/Aspire.Hosting/Backchannel/BackchannelDataTypes.cs
@@ -955,7 +955,6 @@ internal class PublishingPromptInputAnswer
 {
     public string? Name { get; set; }
     public string? Value { get; set; }
-    public string? FileName { get; set; }
 }
 
 internal sealed class UploadFileRequest
